Add files via upload
This dataset contains over 12k observations of male-female baboon pairs from a population of baboons that has recently seen genetic admixture from a different (but closely-related) taxon. The data contains genetic and social information for the male and female baboons, whether they mated, and whether the mating resulted in conception of offspring.